% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{add_logo_gg}
\alias{add_logo_gg}
\title{Add Logo}
\usage{
add_logo_gg(plot, color = style$logo$color)
}
\arguments{
\item{plot}{A ggplot2 plot (of class 'gg' and 'ggplot').}

\item{color}{Character. A color hex.}
}
\value{
A ggplot2 plot object with the logo added.
}
\description{
Adds the Baseline logo to a ggplot2 plot.
}
