% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{add_logo_gt}
\alias{add_logo_gt}
\title{Add \code{gt} Logo}
\usage{
add_logo_gt(table, width = 1000, height = 1000, color = style$logo$color)
}
\arguments{
\item{table}{A magick_image.}

\item{width}{Numeric. A width value in pixels.}

\item{height}{Numeric. A height value in pixels.}

\item{color}{Character. A color hex defaulted to \verb{#333333}.}
}
\description{
Adds the Baseline logo to a \code{gt} table.
}
