% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{baseline_plot}
\alias{baseline_plot}
\title{Baseline Plot}
\usage{
baseline_plot(plot)
}
\arguments{
\item{plot}{A ggplot2 plot object.}
}
\value{
A ggplot2 plot object with the Baseline theme and logo.
}
\description{
Applies the Baseline theme and adds the Baseline logo to a \code{ggplot2} plot.
This is a convenience wrapper which applies the theme using \code{\link{theme_baseline_gg}}
to a plot and then adds the logo to a plot using \code{\link{add_logo_gg}}.
Use the individual functions for more manual control and styling overrides.
}
