% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyregressor.R
\name{dummy_regressor}
\alias{dummy_regressor}
\title{Trains a very basic dummy regressor model.}
\usage{
dummy_regressor(formula, data, strategy, constant, quantile, seed)
}
\arguments{
\item{formula}{that is used during training}

\item{data}{that is used during training}

\item{strategy}{that is applied}
}
\description{
Trains a very basic dummy regressor model.
}
