% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{ggsave_with_logo}
\alias{ggsave_with_logo}
\title{Save \code{ggplot2} Plot with Logo}
\usage{
ggsave_with_logo(plot, output, width = 6, height = 6, dpi = 600)
}
\arguments{
\item{plot}{A \code{ggplot2} plot object.}

\item{output}{A file path where the output image will be saved.}

\item{width}{Numeric. The width of the output image in \strong{inches} defaulted to 6.}

\item{height}{Numeric. The height of the output image in \strong{inches} defaulted to 6.}

\item{dpi}{Numeric. The resolution of the output image in \strong{dots per inch (DPI)} defaulted to 6.}
}
\description{
Adds the Baseline logo to a \code{ggplot2} plot and saves the plot.
}
