% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{gtsave_with_logo}
\alias{gtsave_with_logo}
\title{Save \code{gt} Table with Logo}
\usage{
gtsave_with_logo(table, output, width = 6, height = 6, dpi = 600)
}
\arguments{
\item{table}{A \code{gt} table object.}

\item{output}{A file path where the output image will be saved.}

\item{width}{Numeric. The width of the output image in \strong{inches} defaulted to 6.}

\item{height}{Numeric. The height of the output image in \strong{inches} defaulted to 6.}

\item{dpi}{Numeric. The resolution of the output image in \strong{dots per inch (DPI)} defaulted to 6.}
}
\description{
Adds the Baseline logo to a \code{gt} table and saves the table.
}
