% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanmodel.R
\name{mean_model}
\alias{mean_model}
\title{Trains a very basic mean model.}
\usage{
mean_model(formula, data)
}
\arguments{
\item{formula}{that is used during training}

\item{data}{that is used during training}
}
\description{
Trains a very basic mean model.
}
