% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{theme_baseline_gg}
\alias{theme_baseline_gg}
\title{Baseline Plot Theme}
\usage{
theme_baseline_gg(scale = 1)
}
\arguments{
\item{scale}{Numeric. Scale factor relative to a base plot of 6x6 inches.}
}
\value{
A \code{ggplot2} theme object that can be added to a ggplot.
}
\description{
Apply the Baseline theme to a \code{ggplot2} plot.
}
