% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{theme_baseline_gt}
\alias{theme_baseline_gt}
\title{Baseline Table Theme}
\usage{
theme_baseline_gt(table, ...)
}
\arguments{
\item{table}{An existing gt table object of class \code{gt_tbl}.}

\item{...}{Additional optional arguments passed to \code{gt::tab_options()} to override table styling.}
}
\value{
Returns a styled \code{gt} table.
}
\description{
Apply the Baseline theme to a \code{gt} table.
}
