% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{add_logo_gg}
\alias{add_logo_gg}
\title{Add \code{ggplot2} Logo}
\usage{
add_logo_gg(plot, color = style$logo$color)
}
\arguments{
\item{plot}{A \code{ggplot2} plot object.}

\item{color}{Character. A color hex defaulted to \verb{#333333}.}
}
\value{
A \code{ggplot2} plot object with the logo added.
}
\description{
Adds the Baseline logo to a \code{ggplot2} plot.
}
