% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{theme_baseline_gg}
\alias{theme_baseline_gg}
\title{Baseline Theme (ggplot2)}
\usage{
theme_baseline_gg(scale = 1)
}
\arguments{
\item{scale}{Numeric. Scale factor relative to a base plot of 6x6 inches.}
}
\value{
A ggplot2 theme object that can be added to a ggplot.
}
\description{
Styles the ggplot2 theme for the Baseline.
}
